Address reviewer's comments
