+++ nav2_behavior_tree/include/nav2_behavior_tree/plugins/time_expired_condition.hpp
@@ -13,52 +13,42 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef NAV2_BEHAVIOR_TREE__PLUGINS__CONDITION__DISTANCE_TRAVELED_CONDITION_HPP_
-#define NAV2_BEHAVIOR_TREE__PLUGINS__CONDITION__DISTANCE_TRAVELED_CONDITION_HPP_
+#ifndef NAV2_BEHAVIOR_TREE__PLUGINS__TIME_EXPIRED_CONDITION_HPP_
+#define NAV2_BEHAVIOR_TREE__PLUGINS__TIME_EXPIRED_CONDITION_HPP_
 
 #include <string>
-#include <memory>
-
-#include "behaviortree_cpp_v3/condition_node.h"
 
 #include "rclcpp/rclcpp.hpp"
-#include "geometry_msgs/msg/pose_stamped.hpp"
-#include "tf2_ros/buffer.h"
+#include "behaviortree_cpp_v3/condition_node.h"
 
 namespace nav2_behavior_tree
 {
 
-class DistanceTraveledCondition : public BT::ConditionNode
+class TimeExpiredCondition : public BT::ConditionNode
 {
 public:
-  DistanceTraveledCondition(
+  TimeExpiredCondition(
     const std::string & condition_name,
     const BT::NodeConfiguration & conf);
 
-  DistanceTraveledCondition() = delete;
+  TimeExpiredCondition() = delete;
 
   BT::NodeStatus tick() override;
 
+  // Any BT node that accepts parameters must provide a requiredNodeParameters method
   static BT::PortsList providedPorts()
   {
     return {
-      BT::InputPort<double>("distance", 1.0, "Distance"),
-      BT::InputPort<std::string>("global_frame", std::string("map"), "Global frame"),
-      BT::InputPort<std::string>("robot_base_frame", std::string("base_link"), "Robot base frame")
+      BT::InputPort<double>("seconds", 1.0, "Seconds")
     };
   }
 
 private:
   rclcpp::Node::SharedPtr node_;
-  std::shared_ptr<tf2_ros::Buffer> tf_;
-
-  geometry_msgs::msg::PoseStamped start_pose_;
-
-  double distance_;
-  std::string global_frame_;
-  std::string robot_base_frame_;
+  rclcpp::Time start_;
+  double period_;
 };
 
 }  // namespace nav2_behavior_tree
 
-#endif  // NAV2_BEHAVIOR_TREE__PLUGINS__CONDITION__DISTANCE_TRAVELED_CONDITION_HPP_
+#endif  // NAV2_BEHAVIOR_TREE__PLUGINS__TIME_EXPIRED_CONDITION_HPP_
